polarh10_driver/requirements.txt
bleak>=0.22.0
polar-python>=0.0.5

numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
matplotlib>=3.7.0

fastapi>=0.110.0
uvicorn>=0.29.0

pyyaml>=6.0
websockets==12.0